HttpTriggerEnergyPrices/requirements.txt
selenium
beautifulsoup4
matplotlib
numpy
scipy
cloudinary
twilio
python-dotenv
azure-functions
